// types.ts
export type ResponseType = {
  success: boolean;
  data?: any;
  msg?: string;
};
export type InventoryType = {
  id?: string;
  name: string;
  category: string;
  brand?: string;
  purchasedPrice: number;
  sellingPrice: number;
  expireDate: Date | Timestamp | string;
  quantity?: number;
  description?: string;
  image?: any | null;
  created?: string;
  updated?: string;
  ownerName?: string;
};
export type ExpensesType = {
  id?: string;
  category: string;
  expenseCost: number;
  description?: string;
  image?: any | null;
  ownerName?: string;
  created?: Date | Timestamp | string;
};
// ...
export type SellType = {
  id?:string;
  inventoryProduct: string;
  quantity: number;
  Price: number;
  loan?: boolean;
  loanAmount?: number;
  PriceAfterLoan: number;
  discount?: number;
  date: Date | string | Timestamp;
  customerName?: string;
  customerNumber?: number;
  description?: string;
  ownerName?: string;
};